requirements.txt
pillow~=9.2.0
torchvision~=0.13.1
tqdm~=4.65.0